NumberTwoLabs/requirements.txt
Django==1.7.5
Markdown==2.6
Pillow==2.7.0
django-braces==1.4.0
django-filter==0.9.2
django-oauth-toolkit==0.7.2
django-oauth2-provider==0.2.6.1
djangorestframework==3.0.5
oauthlib==0.7.2
psycopg2==2.6
shortuuid==0.4.2
six==1.9.0
wsgiref==0.1.2
